update for gcc3
git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/vizmo/trunk@68 8e255c04-9930-0410-9bf1-9572f6c5958d

+++ vizmo2/src/EnvObj/PathLoader.cpp
@@ -56,7 +56,7 @@ bool CPathLoader::ParseFile()
 
 double * CPathLoader::GetConfiguration(int frame)
 {
-	if(frame>=m_pList.size()) frame=m_pList.size()-1;
+	if(frame>=(int)m_pList.size()) frame=m_pList.size()-1;
 	if(frame<0 ) frame=0;
 		
     double * currentCfg = GetConfigure(false, frame);
